requirements.txt
fastapi
uvicorn
yfinance
pandas
requests
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
google-genai
jinja2
python-multipart